pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "mss==9.0.1",
    "numpy==1.26.3",
    "opencv-python==4.9.0.80",
    "pillow>=12.0.0",
    "playwright==1.41.0",
    "psutil>=7.2.1",
    "pyautogui>=0.9.54",
    "pynput==1.7.6",
    "pyobjc-framework-cocoa==10.1 ; sys_platform == 'darwin'",
    "python-dotenv>=1.2.1",
    "pywin32>=306 ; sys_platform == 'win32'",
    "requests>=2.32.5",
]
